+++ src/main/scala/metrics/Metrics.scala
@@ -1,11 +1,13 @@
 package metrics
 
-import breeze.linalg.{DenseMatrix, DenseVector, max, min}
+import breeze.linalg.{norm, _}
+import breeze.math._
+import breeze.numerics._
 import cluster.Types
 
 object Metrics {
 
-  def par(e1: DenseVector[Int], e2: DenseVector[Int]): Double = {
+  def par(e1: DenseVector[Double], e2: DenseVector[Double]): Double = {
     val sum = e1 + e2
     max(sum) - min(sum)
   }
